Added: +Vector length calculations +Vector normalization calculations +Image rotations
Changed:
*Floating point based movement for sprites

+++ project/src/call/game/entitys/BasicEntity.java
@@ -71,12 +71,12 @@ public class BasicEntity
 		setHealth(0);
 	}
 
-	public int getX()
+	public double getX()
 	{
 		return this.sprite.getX();
 	}
 
-	public int getY()
+	public double getY()
 	{
 		return this.sprite.getY();
 	}
